+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/API Bindings/Link/STPAPIClient+Link.swift
@@ -66,7 +66,7 @@ extension STPAPIClient {
                     // If there's an assertion error, send it to StripeAttest
                     if useMobileEndpoints,
                        case .failure(let error) = result,
-                       Self.isLinkAssertionError(error: error) {
+                       StripeAttest.isLinkAssertionError(error: error) {
                         await self.stripeAttest.receivedAssertionError(error)
                     }
                     // Mark the assertion handle as completed
@@ -132,7 +132,7 @@ extension STPAPIClient {
                     // If there's an assertion error, send it to StripeAttest
                     if useMobileEndpoints,
                        case .failure(let error) = result,
-                       Self.isLinkAssertionError(error: error) {
+                       StripeAttest.isLinkAssertionError(error: error) {
                         await self.stripeAttest.receivedAssertionError(error)
                     }
                     // Mark the assertion handle as completed
@@ -141,15 +141,6 @@ extension STPAPIClient {
                 }
             }
         }
-    }
-
-    @_spi(STP) public static func isLinkAssertionError(error: Error) -> Bool {
-        if let error = error as? StripeCore.StripeError,
-           case let .apiError(apiError) = error,
-           apiError.code == "link_failed_to_attest_request" {
-            return true
-        }
-        return false
     }
 
     private func makePaymentDetailsRequest(